NULL-536-feat: 메모 정리 분산처리락 구현
