--- src/main/java/com/example/oatnote/domain/memotag/service/AsyncMemoTagService.java
@@ -1,12 +1,0 @@
-package com.example.oatnote.domain.memotag.service;
-
-import org.springframework.stereotype.Service;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class AsyncMemoTagService {
-
-
-}
